Fix crash for the flashSettings observer in CameraViewFragment
On a brand new install, setting the observer for flashSettings in the CameraViewFragment will cause the application to crash. This occurs because the observer tries to modify the CameraView object before the application has asked permission from the user to use the Camera. After permissions has been granted the CameraView object will then be initialized.

The solution was to create the observer for flashSettings after the user has granted permissions and after checking for permissions. Therefore the function initCameraObservers() was created to separate any Observers for specifically modifying the CameraView object.