Fix the Favorites filter where highlighted images were not deselected
When the user cancels the deletion selection on the ThumbnailGalleryFragment some thumbnails were still being highlighted. This was due to the onItemRangeInserted() AdapterDataObserver callback repositioning the views incorrectly. Initially this method was overridden to pull the recycler view to the top when new items were inserted. The body of the function call back was re-modified to prevent unnecessary modifications to the insertion of the items.