Fix negative indexing when hearting/un-hearting an photo image
Problem: If the user toggles the heart button in the MediaViewerFragment and tries to navigate back again from the thumbnail gallery the MediaViewerFragment will show the very first photo. This is caused by the position being calculated by the indexOf function in the ImageFile list. The heart property in the ImageFile is changed which in turn changes the data of the object. This will cause the indexOf function to return a -1.

Solution: To fix this, the function indexOfFirst() is used with the predicate comparing the file path which is a constant and will never change.